Move testing utils to `pyoptsparse.testing`
