+++ pyoptsparse/testing/__init__.py
@@ -1,0 +1,1 @@
+from .pyOpt_testing import *
