+++ tests/test_hs071.py
@@ -10,9 +10,7 @@ from parameterized import parameterized
 
 # First party modules
 from pyoptsparse import History, Optimization
-
-# Local modules
-from testing_utils import OptTest
+from pyoptsparse.testing import OptTest
 
 
 class TestHS71(OptTest):
