+++ tests/test_nsga2_multi_objective.py
@@ -9,9 +9,7 @@ from parameterized import parameterized
 
 # First party modules
 from pyoptsparse import Optimization
-
-# Local modules
-from testing_utils import OptTest
+from pyoptsparse.testing import OptTest
 
 
 class TestNSGA2(OptTest):
