+++ tests/test_optProb.py
@@ -12,7 +12,7 @@ from numpy.testing import assert_allclose
 from pyoptsparse import OPT, Optimization
 
 # Local modules
-from testing_utils import assert_optProb_size
+from pyoptsparse.testing.pyOpt_testing import assert_optProb_size
 
 
 class TestOptProb(unittest.TestCase):
