+++ tests/test_tp109.py
@@ -37,9 +37,7 @@ from parameterized import parameterized
 
 # First party modules
 from pyoptsparse import History, Optimization
-
-# Local modules
-from testing_utils import OptTest
+from pyoptsparse.testing import OptTest
 
 USE_LINEAR = True
